@@ -11,16 +11,13 @@ use Symfony\Component\Routing\Annotation\Route;
 
 final class PostController extends AbstractController
 {
-    private PostRepository $postRepository;
-
-    public function __construct(PostRepository $postRepository)
+    public function __construct(private PostRepository $postRepository)
     {
-        $this->postRepository = $postRepository;
     }
 
-    /**
-     * @Route(path="blog/{postSlug}", name="post", requirements={"postSlug"=".+"})
-     */
+    #[Route('blog/{postSlug}', name: 'post', requirements: [
+        'postSlug' => '.+',
+    ])]
     public function __invoke(string $postSlug): Response
     {
         $post = $this->postRepository->findBySlug($postSlug);
@@ -28,7 +25,6 @@ final class PostController extends AbstractController
             $message = sprintf("Post with slug '%s' not found", $postSlug);
             throw $this->createNotFoundException($message);
         }
-
         return $this->render('blog/post.twig', [
             'post' => $post,
         ]);
