@@ -1,13 +1,35 @@
 <?php
 
+declare(strict_types=1);
+
 return [
-    Symfony\Bundle\FrameworkBundle\FrameworkBundle::class => ['all' => true],
-    Symfony\Bundle\DebugBundle\DebugBundle::class => ['dev' => true, 'test' => true],
-    Symfony\Bundle\TwigBundle\TwigBundle::class => ['all' => true],
-    Symfony\Bundle\WebProfilerBundle\WebProfilerBundle::class => ['dev' => true, 'test' => true],
-    Symfony\Bundle\SecurityBundle\SecurityBundle::class => ['all' => true],
-    Doctrine\Bundle\DoctrineBundle\DoctrineBundle::class => ['all' => true],
-    Doctrine\Bundle\MigrationsBundle\DoctrineMigrationsBundle::class => ['all' => true],
-    Symfony\WebpackEncoreBundle\WebpackEncoreBundle::class => ['all' => true],
-    Sentry\SentryBundle\SentryBundle::class => ['all' => true],
+    Symfony\Bundle\FrameworkBundle\FrameworkBundle::class => [
+        'all' => true,
+    ],
+    Symfony\Bundle\DebugBundle\DebugBundle::class => [
+        'dev' => true,
+        'test' => true,
+    ],
+    Symfony\Bundle\TwigBundle\TwigBundle::class => [
+        'all' => true,
+    ],
+    Symfony\Bundle\WebProfilerBundle\WebProfilerBundle::class => [
+        'dev' => true,
+        'test' => true,
+    ],
+    Symfony\Bundle\SecurityBundle\SecurityBundle::class => [
+        'all' => true,
+    ],
+    Doctrine\Bundle\DoctrineBundle\DoctrineBundle::class => [
+        'all' => true,
+    ],
+    Doctrine\Bundle\MigrationsBundle\DoctrineMigrationsBundle::class => [
+        'all' => true,
+    ],
+    Symfony\WebpackEncoreBundle\WebpackEncoreBundle::class => [
+        'all' => true,
+    ],
+    Sentry\SentryBundle\SentryBundle::class => [
+        'all' => true,
+    ],
 ];
