@@ -11,11 +11,8 @@ use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 final class RectorRunRepository
 {
-    private EntityManagerInterface $entityManager;
-
-    public function __construct(EntityManagerInterface $entityManager)
+    public function __construct(private EntityManagerInterface $entityManager)
     {
-        $this->entityManager = $entityManager;
     }
 
     public function save(RectorRun $rectorRun): void
