@@ -4,14 +4,11 @@ declare(strict_types=1);
 
 namespace Rector\Website\Demo\Validator\Constraint;
 
-use Doctrine\Common\Annotations\Annotation\Target;
+use Attribute;
 use Rector\Website\Demo\Validator\PHPConstraintValidator;
 use Symfony\Component\Validator\Constraint;
 
-/**
- * @Annotation
- * @Target({"PROPERTY", "ANNOTATION"})
- */
+#[Attribute(Attribute::TARGET_PROPERTY)]
 final class PHPConstraint extends Constraint
 {
     public function validatedBy(): string
