@@ -8,20 +8,12 @@ use Rector\Website\Demo\Validator\Constraint\PHPConstraint;
 
 final class DemoFormData
 {
-    /**
-     * @PHPConstraint()
-     */
-    private string $content;
-
-    /**
-     * @PHPConstraint()
-     */
-    private string $config;
-
-    public function __construct(string $content, string $config)
-    {
-        $this->content = $content;
-        $this->config = $config;
+    public function __construct(
+        #[PHPConstraint]
+        private string $content,
+        #[PHPConstraint]
+        private string $config,
+    ) {
     }
 
     public function getContent(): string
